# docs/swagger.yaml
definitions:
  cwrs_res.ResError:
    properties:
      code:
        type: integer
      errMsg:
        type: string
      msg:
        type: string
    type: object
  cwrs_res.ResSuccess:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  cwrs_res.ResSuccessData:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  cwrs_res.ResSuccessDataList:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      total:
        type: integer
    type: object
...
  pojo.AddSysRoleReq:
    description: 新增角色
    properties:
      deptId:
        description: 所属组织
        type: string
      desc:
        description: 描述
        type: string
      roleCode:
        description: 角色编码
        type: string
      roleMenus:
        description: 平台菜单
        items:
          $ref: '#/definitions/pojo.RoleMenuAddReq'
        type: array
      roleMenusApp:
        description: APP菜单
        items:
          $ref: '#/definitions/pojo.RoleMenuAddReq'
        type: array
      roleName:
        description: 角色名称
        type: string
      roleSort:
        description: 排序
        type: integer
      roleStatus:
        description: 状态(字典-sys_status)
        type: string
    required:
    - deptId
    - roleCode
    - roleName
    - roleSort
    - roleStatus
    type: object
  pojo.AddSysTaskReq:
    properties:
      cronName:
        description: 执行时间(字典-sys_task_cron_expr)
        type: string
      desc:
        description: 描述
        type: string
      funcName:
        description: 要执行的函数名
        type: string
      taskName:
        description: 任务名称
        type: string
      taskParams:
        allOf:
        - $ref: '#/definitions/pojo.JSONMap'
        description: JSON格式参数
    required:
    - cronName
    - funcName
    - taskName
    type: object
...
  pojo.EditSysRoleReq:
    description: 修改角色
    properties:
      deptId:
        description: 所属组织
        type: string
      desc:
        description: 描述
        type: string
      roleCode:
        description: 角色编码
        type: string
      roleId:
        description: 角色id
        type: string
      roleMenus:
        description: 平台菜单
        items:
          $ref: '#/definitions/pojo.RoleMenuAddReq'
        type: array
      roleMenusApp:
        description: APP菜单
        items:
          $ref: '#/definitions/pojo.RoleMenuAddReq'
        type: array
      roleName:
        description: 角色名称
        type: string
      roleSort:
        description: 排序
        type: integer
      roleStatus:
        description: 状态(字典-sys_status)
        type: string
    required:
    - deptId
    - roleCode
    - roleId
    - roleName
    - roleSort
    - roleStatus
    type: object
  pojo.EditSysTaskReq:
    properties:
      cronName:
        description: 执行时间(字典-sys_task_cron_expr)
        type: string
      desc:
        description: 描述
        type: string
      funcName:
        description: 要执行的函数名
        type: string
      taskId:
        description: 任务ID
        type: string
      taskName:
        description: 任务名称
        type: string
      taskParams:
        allOf:
        - $ref: '#/definitions/pojo.JSONMap'
        description: JSON格式参数
    required:
    - cronName
    - funcName
    - taskId
    - taskName
    type: object
...
  pojo.JSONMap:
    additionalProperties: true
    type: object
  pojo.RoleMenuAddReq:
    properties:
      dataPurview:
        description: 数据权限(字典-sys_data_purview)-按菜单控制
        type: string
      deptIds:
        description: 自定义数据权限授权组织ids
        items:
          type: string
        type: array
      menuId:
        description: 菜单id
        type: string
      menuType:
        description: 菜单类型
        type: string
    type: object
  pojo.RoleMenuResp:
    properties:
      dataPurview:
        description: 数据权限(字典-sys_data_purview)-按菜单控制
        type: string
      deptIds:
        description: 自定义数据权限授权组织ids
        items:
          type: string
        type: array
      menuId:
        description: 菜单id
        type: string
      menuType:
        description: 菜单类型
        type: string
    type: object
...
  pojo.SysOperLogListResp:
    properties:
      ip:
        description: 客户端IP
        type: string
      latency:
        description: 耗时
        type: string
      method:
        description: 请求方法
        type: string
      operId:
        description: 日志主键
        type: string
      operName:
        description: 操作用户名称
        type: string
      operTime:
        description: 操作时间
        type: string
      operUserId:
        description: 操作用户id
        type: string
      path:
        description: 请求路径
        type: string
      status:
        description: 响应状态码
        type: integer
    type: object
  pojo.SysOperLogResp:
    properties:
      ip:
        description: 客户端IP
        type: string
      latency:
        description: 耗时
        type: string
      method:
        description: 请求方法
        type: string
      operId:
        description: 日志主键
        type: string
      operName:
        description: 操作用户名称
        type: string
      operTime:
        description: 操作时间
        type: string
      operUserId:
        description: 操作用户id
        type: string
      path:
        description: 请求路径
        type: string
      reqBody:
        description: 请求参数
        type: string
      resBody:
        description: 响应数据
        type: string
      status:
        description: 响应状态码
        type: integer
    type: object
...
  pojo.SysRoleDetailResp:
    properties:
      createdTime:
        description: 创建时间
        type: string
      createdUserId:
        description: 创建用户
        type: string
      createdUserName:
        description: 创建用户名称
        type: string
      deptId:
        description: 所属组织
        type: string
      deptName:
        description: 所属组织名称
        type: string
      desc:
        description: 描述
        type: string
      isBuiltin:
        description: 是否内置1是 0否
        type: integer
      roleCode:
        description: 角色编码
        type: string
      roleId:
        description: 主键
        type: string
      roleMenus:
        description: 平台菜单
        items:
          $ref: '#/definitions/pojo.RoleMenuResp'
        type: array
      roleMenusApp:
        description: APP菜单
        items:
          $ref: '#/definitions/pojo.RoleMenuResp'
        type: array
      roleName:
        description: 角色名称
        type: string
      roleSort:
        description: 排序 从1开始
        type: integer
      roleStatus:
        description: 状态(字典-sys_status)
        type: string
    type: object
  pojo.SysRoleResp:
    properties:
      createdTime:
        description: 创建时间
        type: string
      createdUserId:
        description: 创建用户
        type: string
      createdUserName:
        description: 创建用户名称
        type: string
      deptId:
        description: 所属组织
        type: string
      deptName:
        description: 所属组织名称
        type: string
      desc:
        description: 描述
        type: string
      isBuiltin:
        description: 是否内置1是 0否
        type: integer
      roleCode:
        description: 角色编码
        type: string
      roleId:
        description: 主键
        type: string
      roleName:
        description: 角色名称
        type: string
      roleSort:
        description: 排序 从1开始
        type: integer
      roleStatus:
        description: 状态(字典-sys_status)
        type: string
    type: object
  pojo.SysTaskResp:
    properties:
      createdTime:
        description: 创建时间
        type: string
      createdUserId:
        description: 创建用户
        type: string
      createdUserName:
        description: 创建用户名称
        type: string
      cronExpr:
        description: Cron表达式-字典值
        type: string
      cronName:
        description: 执行时间(字典-sys_task_cron_expr)
        type: string
      desc:
        description: 描述
        type: string
      funcName:
        description: 要执行的函数名
        type: string
      lastRunTime:
        description: 最近一次开启/关闭时间
        type: string
      taskId:
        description: 任务ID
        type: string
      taskName:
        description: 任务名称
        type: string
      taskParams:
        allOf:
        - $ref: '#/definitions/pojo.JSONMap'
        description: JSON格式参数
      taskStatus:
        description: 任务状态(字典-sys_task_status)
        type: string
    type: object
...
host: localhost:9090
info:
  contact:
    email: vn_wang@163.com
    name: vn_wang
  description: Cwrs管理系统API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: 本系统为开源项目，可用于学习和研究，如用于商业用途造成的后果与本组织无关。
  title: Cwrs Manager API
  version: "1.0"
paths:
...
  /sysOperLog/detail:
    get:
      consumes:
      - application/json
      description: 根据主键查询操作日志详情
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: id
        in: query
        name: operId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            allOf:
            - $ref: '#/definitions/cwrs_res.ResSuccessData'
            - properties:
                data:
                  $ref: '#/definitions/pojo.SysOperLogResp'
              type: object
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 查询操作日志详情
      tags:
      - 操作日志管理【平台】
  /sysOperLog/list:
    get:
      consumes:
      - application/json
      description: 分页查询操作日志列表
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 操作时间结束
        in: query
        name: endTime
        type: string
      - description: 操作用户名称
        in: query
        name: operName
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页显示条数
        in: query
        name: pageSize
        type: integer
      - description: 操作时间开始
        in: query
        name: startTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            allOf:
            - $ref: '#/definitions/cwrs_res.ResSuccessDataList'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pojo.SysOperLogListResp'
                  type: array
              type: object
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 分页查询操作日志列表
      tags:
      - 操作日志管理【平台】
...
  /sysRole/add:
    post:
      consumes:
      - application/json
      description: 新增角色
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 新增角色参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/pojo.AddSysRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/cwrs_res.ResSuccess'
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 新增角色
      tags:
      - 角色管理【平台】
  /sysRole/del:
    delete:
      consumes:
      - application/json
      description: 批量删除角色
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 角色id 多个英文逗号分隔
        in: query
        name: roleIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/cwrs_res.ResSuccess'
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 删除角色
      tags:
      - 角色管理【平台】
  /sysRole/detail:
    get:
      consumes:
      - application/json
      description: 根据主键查询角色详情
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 角色id
        in: query
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            allOf:
            - $ref: '#/definitions/cwrs_res.ResSuccessData'
            - properties:
                data:
                  $ref: '#/definitions/pojo.SysRoleDetailResp'
              type: object
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 查询角色详情
      tags:
      - 角色管理【平台】
  /sysRole/edit:
    put:
      consumes:
      - application/json
      description: 修改角色
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 修改角色参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/pojo.EditSysRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/cwrs_res.ResSuccess'
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 修改角色
      tags:
      - 角色管理【平台】
  /sysRole/list:
    get:
      consumes:
      - application/json
      description: 分页查询角色列表
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 所属组织
        in: query
        name: deptId
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页显示条数
        in: query
        name: pageSize
        type: integer
      - description: 角色编码(模糊)
        in: query
        name: roleCode
        type: string
      - description: 角色名称(模糊)
        in: query
        name: roleName
        type: string
      - description: 状态
        in: query
        name: roleStatus
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            allOf:
            - $ref: '#/definitions/cwrs_res.ResSuccessDataList'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pojo.SysRoleResp'
                  type: array
              type: object
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 分页查询角色列表
      tags:
      - 角色管理【平台】
  /sysTask/add:
    post:
      consumes:
      - application/json
      description: 新增定时任务
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 新增参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/pojo.AddSysTaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/cwrs_res.ResSuccess'
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 新增定时任务
      tags:
      - 定时任务管理【平台】
  /sysTask/del:
    delete:
      consumes:
      - application/json
      description: 删除定时任务 只能删除停止状态的任务
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: id 多个英文逗号分隔
        in: query
        name: taskIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/cwrs_res.ResSuccess'
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 删除定时任务 只能删除停止状态的任务
      tags:
      - 定时任务管理【平台】
  /sysTask/detail:
    get:
      consumes:
      - application/json
      description: 根据主键查询定时任务详情
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: id
        in: query
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            allOf:
            - $ref: '#/definitions/cwrs_res.ResSuccessData'
            - properties:
                data:
                  $ref: '#/definitions/pojo.SysTaskResp'
              type: object
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 查询定时任务详情
      tags:
      - 定时任务管理【平台】
  /sysTask/edit:
    put:
      consumes:
      - application/json
      description: 修改定时任务 只能修改停止状态的任务
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 修改参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/pojo.EditSysTaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/cwrs_res.ResSuccess'
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 修改定时任务 只能修改停止状态的任务
      tags:
      - 定时任务管理【平台】
  /sysTask/list:
    get:
      consumes:
      - application/json
      description: 分页查询定时任务列表
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页显示条数
        in: query
        name: pageSize
        type: integer
      - description: 任务名称
        in: query
        name: taskName
        type: string
      - description: 任务状态(字典-sys_task_status)
        in: query
        name: taskStatus
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            allOf:
            - $ref: '#/definitions/cwrs_res.ResSuccessDataList'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pojo.SysTaskResp'
                  type: array
              type: object
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 分页查询定时任务列表
      tags:
      - 定时任务管理【平台】
  /sysTask/start:
    get:
      consumes:
      - application/json
      description: 打开定时任务
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: id 多个英文逗号分隔
        in: query
        name: taskIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/cwrs_res.ResSuccess'
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 打开定时任务
      tags:
      - 定时任务管理【平台】
  /sysTask/stop:
    get:
      consumes:
      - application/json
      description: 关闭定时任务
      parameters:
      - description: token
        in: header
        name: Token
        required: true
        type: string
      - description: id 多个英文逗号分隔
        in: query
        name: taskIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/cwrs_res.ResSuccess'
        "422":
          description: 操作失败
          schema:
            $ref: '#/definitions/cwrs_res.ResError'
      summary: 关闭定时任务
      tags:
      - 定时任务管理【平台】
...
swagger: "2.0"